update(admin): UI

+++ admin/categories/index.php
@@ -75,7 +75,7 @@
                                         </a>
                                     </td>
                                     <td>
-                                        <a href="delete.php?id=<?= $each['id'] ?>">
+                                        <a onclick="return confirm('Bạn chắc chắn muốn xóa?')" href="delete.php?id=<?= $each['id'] ?>">
                                         <i class="bi bi-trash-fill"></i>
                                         </a>
                                     </td>
